requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2==2.9.10
python-dotenv==1.0.1